% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{airports}
\alias{airports}
\title{airports}
\format{A 3,376 x 7 dataframe listing the location of each airport in
the U.S.A}
\source{
I got this from Hadley Wickam's *Data Manipulation with dplyr*
presentation from June 2014
}
\usage{
airports
}
\description{
Airport location metadata
}
\keyword{datasets}
