% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{beatles_green}
\alias{beatles_green}
\title{beatles_green}
\format{A 5 x 2 dataframe of meaningless data to demonstrating basic
*dplyr* join functions based on the Beatles}
\source{
I got this from Hadley Wickam's *Data Manipulation with dplyr*
 presentation from June 2014
}
\usage{
beatles_green
}
\description{
A simple example dataset for demonstrating basic *dplyr* functions
}
\keyword{datasets}
