% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{This function is a conversation starter.}
\usage{
hi(who, how = c("shout", "whisper"))
}
\arguments{
\item{who}{character(1) The name of the person for conversation}

\item{how}{character(1) How to greet the conversant. "shout" or "whisper".}
}
\value{
character(1) conversation starter
}
\description{
some helpful text.
}
\examples{
   hi("Qian Liu")
   hi("Qian Liu", "whisper")
}

