% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{life_expectancies}
\alias{life_expectancies}
\title{life_expectancies}
\format{A 142 x 8 dataframe life expectency data for each country from 1957-2007.}
\source{
I got this from the Gapminder Foundation \url(http://www.gapminder.org/)
via Jenny Bryan's gapminder package \url(https://github.com/jennybc/gapminder)
}
\usage{
life_expectancies
}
\description{
A simple example dataset for demonstrating basic *dplyr* functions
}
\keyword{datasets}
