% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{planes}
\alias{planes}
\title{planes}
\format{A 2853 x 9 dataframe listing the technical capacities of
the planes which flwe out of Houston in 2011}
\source{
I got this from Hadley Wickam's *Data Manipulation with dplyr*
 presentation from June 2014
}
\usage{
planes
}
\description{
Plane metadata
}
\keyword{datasets}
