% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\docType{data}
\name{weather}
\alias{weather}
\title{weather}
\format{A 8723 x 16 dataframe listing the weather data for Houston for
the year 2011.}
\source{
I got this from Hadley Wickam's *Data Manipulation with dplyr*
 presentation from June 2014
}
\usage{
weather
}
\description{
Hourly weather data for Houston for the year 2011
}
\keyword{datasets}
